+++ mobx_examples/lib/github/github_store.g.dart
@@ -25,8 +25,7 @@ mixin _$GithubStore on _GithubStore, Store {
 
   @override
   set fetchReposFuture(ObservableFuture<List<Repository>> value) {
-    _$fetchReposFutureAtom.context
-        .checkIfStateModificationsAreAllowed(_$fetchReposFutureAtom);
+    _$fetchReposFutureAtom.context.enforceWritePolicy(_$fetchReposFutureAtom);
     super.fetchReposFuture = value;
     _$fetchReposFutureAtom.reportChanged();
   }
@@ -41,7 +40,7 @@ mixin _$GithubStore on _GithubStore, Store {
 
   @override
   set user(String value) {
-    _$userAtom.context.checkIfStateModificationsAreAllowed(_$userAtom);
+    _$userAtom.context.enforceWritePolicy(_$userAtom);
     super.user = value;
     _$userAtom.reportChanged();
   }